Support for float service time value

+++ accessmon.py
@@ -268,11 +268,11 @@ def foreverParse(rewind, printBad, logfilepattern, columns):
             svctime = False
             try:
                 if 'times' in columns:
-                    svctime = int(loginfos[columns.get('times')])
+                    svctime = float(loginfos[columns.get('times')])
                 elif 'timems' in columns:
-                    svctime = int(loginfos[columns.get('timems')]) / 1E3
+                    svctime = float(loginfos[columns.get('timems')]) / 1E3
                 elif 'timeus' in columns:
-                    svctime = int(loginfos[columns.get('timeus')]) / 1E6
+                    svctime = float(loginfos[columns.get('timeus')]) / 1E6
             except:
                 badLine = True
